Consolidate old and new CDSR controllers with version mapping
